Get rid of MySQL to simplify the storage, use better schema typing

+++ src/hackernews/data_types.rs
@@ -2,7 +2,7 @@ use crate::DigestItem;
 use serde::Deserialize;
 
 /// A news item that has been fetched from the API
-#[derive(Debug, Deserialize)]
+#[derive(Deserialize)]
 pub struct JsonNewsItem {
     id: i64,
     title: Option<String>,
@@ -14,10 +14,10 @@ impl JsonNewsItem {
     /// Convert to a `DigestItem` for storage
     pub fn as_digest_item(&self) -> DigestItem {
         let mut item = DigestItem {
-            id: self.id as i32,
+            id: self.id,
             news_title: self.title.clone().unwrap_or_default(),
             news_url: self.url.clone().unwrap_or_default(),
-            created_at: self.time as i32,
+            created_at: self.time,
         };
 
         if item.news_url.is_empty() {
